+++ src/OpenTelemetry.Exporter.Geneva/Metrics/TlvMetricExporter.cs
@@ -646,8 +646,8 @@ internal sealed class TlvMetricExporter : IDisposable
                 // TODO: Data Validation
             }
 
-            if (tag.Key.Equals(GenevaMetricExporter.DimensionKeyForCustomMonitoringAccount, StringComparison.OrdinalIgnoreCase) ||
-                tag.Key.Equals(GenevaMetricExporter.DimensionKeyForCustomMetricsNamespace, StringComparison.OrdinalIgnoreCase))
+            if (string.Equals(tag.Key, GenevaMetricExporter.DimensionKeyForCustomMonitoringAccount, StringComparison.OrdinalIgnoreCase) ||
+                string.Equals(tag.Key, GenevaMetricExporter.DimensionKeyForCustomMetricsNamespace, StringComparison.OrdinalIgnoreCase))
             {
                 reservedTags++;
                 continue;
@@ -667,7 +667,7 @@ internal sealed class TlvMetricExporter : IDisposable
         // Serialize MetricPoint Dimension values
         foreach (var tag in tags)
         {
-            if (tag.Key.Equals(GenevaMetricExporter.DimensionKeyForCustomMonitoringAccount, StringComparison.OrdinalIgnoreCase) && tag.Value is string metricsAccount)
+            if (string.Equals(tag.Key, GenevaMetricExporter.DimensionKeyForCustomMonitoringAccount, StringComparison.OrdinalIgnoreCase) && tag.Value is string metricsAccount)
             {
                 if (!string.IsNullOrWhiteSpace(metricsAccount))
                 {
@@ -677,7 +677,7 @@ internal sealed class TlvMetricExporter : IDisposable
                 continue;
             }
 
-            if (tag.Key.Equals(GenevaMetricExporter.DimensionKeyForCustomMetricsNamespace, StringComparison.OrdinalIgnoreCase) && tag.Value is string metricsNamespace)
+            if (string.Equals(tag.Key, GenevaMetricExporter.DimensionKeyForCustomMetricsNamespace, StringComparison.OrdinalIgnoreCase) && tag.Value is string metricsNamespace)
             {
                 if (!string.IsNullOrWhiteSpace(metricsNamespace))
                 {