+++ src/OpenTelemetry.Resources.Process/ProcessSemanticConventions.cs
@@ -5,6 +5,7 @@ namespace OpenTelemetry.Resources.Process;
 
 internal static class ProcessSemanticConventions
 {
+    public const string AttributeProcessCreationTime = "process.creation.time";
     public const string AttributeProcessOwner = "process.owner";
     public const string AttributeProcessPid = "process.pid";
 }