+++ test/OpenTelemetry.Instrumentation.ConfluentKafka.Tests/KafkaCollection.cs
@@ -1,0 +1,14 @@
+// Copyright The OpenTelemetry Authors
+// SPDX-License-Identifier: Apache-2.0
+
+using Xunit;
+
+namespace OpenTelemetry.Instrumentation.ConfluentKafka.Tests;
+
+[CollectionDefinition(Name)]
+#pragma warning disable CA1711 // Identifiers should not have incorrect suffix
+public sealed class KafkaCollection : ICollectionFixture<KafkaFixture>
+#pragma warning restore CA1711 // Identifiers should not have incorrect suffix
+{
+    public const string Name = "Kafka";
+}